backend/requirements.txt
fastapi==0.124.4
uvicorn[standard]==0.33.0
sqlalchemy==2.0.48
aiosqlite==0.20.0
arxiv==2.3.2
httpx==0.28.1
apscheduler==3.11.2
pydantic==2.10.6
python-dotenv==1.0.1
